Various improvements to SmoothScroll.m > eventTapCallback.
- Removed dipatch_async because it led to crashes for some reason.
Replaced it with a check whether the mouse moved in the case of
setDisplayLinkToDisplayUnderMouse().
- optimizations in the form of not doing the scrollswipe recognition
block when smooth scrolling is off. Use mouseMoved instead of scroll
swipes in order to prevent excessive setConfigVariablesForActiveApp()
calls (that function is enormously slow for some reason)
- implemented proper checking for whether the scroll direction has
changed, and added resetting some MomentumPhase variables when the
scroll direction changes, to prevent very slow scrolling when quickly
changing scroll direction.