Revert "Removed Command-Shift to scroll horizontally disabling button in Override Panel" This reverts commit 274da0ef135621c3b1cbdaeb0ccb7f60c7b8b46a.
Found out that rough scroll doesn't actually scroll horizontally by itself in the safari tab bar when pressing shift, for some reason. This is really weird as all that Rough Scroll does is invert the event and post it back into the event stream. So we reverted the commit which removed the horizontal scroll modifier disabling from the app specific settings.