Lot's of refactoring and cleanup in HelperServices.m
Before refactoring HelperServices we had to refactor lots of other stuff.
HelperServices.m came from the MainApp but at some point we copied it to the Helper as a quick hack, cause Helper needed a lot of capabilites from HelperServices.m for deinstallation.
But since we now need to refactor HelperServices to work with the new App architecture instead of the old Prefpane architecture, we now cleaned this up and made HelperServices a shared class between Helper and MainApp to prevent duplication.

HelperServices depends on knowing Helper and MainApp AppBundles. And since the mechanisms for getting app bundles differs slightly for MainApp and Helper, we created another SharedClass called HelperServices which provides AppBundles. Then we refactored other parts of the app to use this new class too.

Refactored some more stuff which depended on hardcoded strings which don't hold anymore under the new architecture. Instead we use values defined in the Shared Class `Constants.h` whereever possible