Made the app compileable on arm. There was a library for zipping files which I depended on for the automatic app update mechanism called ZipArchive. The old version of the framework which I used wasn't buit for arm so I couldn't build the whole app for arm.
The error messages when using Carthage to update the framework were super cyptic and I didn't know what was going on, but now I figured out that the framework dropped support for Carthage, and how to build and include the framework in my project myself. So now the app can be built for arm, yay!

I also removed Carthage from the project, cause that was the only thing I needed it for.