Removed the toolbar, and the switches and the round ellipses more button. And went back to the old checkbox for enabled and push button for more panel.
Using a switch from GitHub to use on older macOS versions is to finicky and prone to breaking in future releases, also the one I tried was nice but didn't support dark mode out of the box. The other one I saw used swift, and I'm not sure it's worth the hassle to include that and have the bundle size quitupled and all that.

The ellipses more button I made also looked a bit iffy when using darkmode. I will stick to non-custom UI elements from now on.