Minor refactoring.
Capturing modified drag doesn't work.
I feel like it might have to do with some optimizations we made because we assumed that modfied drags would always have to depend on a button click.
I recently noticed that a keyboard-only trigger doesn't work to trigger modified drag either, and I feel it might have to do with that.
So I might try to get the keyboard only thing working and hope this clears itself out along with that.