We actually made a critical mistake in our last refactor. (As sort of expected)
We used to pass in `activeModifiers` and `remaps` as well as `effectiveRemaps` to `assessMappingLandscapeWithButton:`. That confused us because `effectiveRemaps` should be determined by `activeModifiers` and `remaps`. Which is part of why we did the refactor.

So we removed `effectiveRemaps` from the function parameters, and we passed in `activeModifiersFiltered` instead. because we thought that we could infer `effectiveRemaps` within the function based on that.

Butttt we missed, that there is another piece to the puzzle of calculating the old `effectiveRemaps` value (which we used to pass in directly) based on the the new method parameters. And that is, what we are now calling the `MFEffectiveRemapsMethod`. We didn't use a simple dictionary lookup to get `effectiveRemaps` from `remaps` and `activeModifiers`, but instead we called a special function which first took the default Modifications (the ones without any preconditions) and then applied the Modifications with a precondition of `activeModifiers` as an override to those default ones.

After the refactor in the last commit, this was lost and we just performed a simple dictionary lookup in the refactored `assessMappingLandscapeWithButton:`. Now we're passing in a block instead containinng the method that should be used.

A design goal with `assessMappingLandscapeWithButton:` was that it's a pure function, so it's result should purely depend on the arguments chosen. I thought that might make it less error prone, which I thought was important, because this stuff is quite confusing.
The way the function arguments were set up before actually made sense in that it was reusing information we calculated before maximally, while still being a pure function, but it was also really hard to understand what it does exactly and what the parameters should be. This lead to issues, because I reused the function elsewhere (in `buttonCouldStillBeUsedThisClickCycle:`) and passed in wrong parameters. I actually adjusted the parameters at some point and it was still wrong. So it was evidently very confusing.

With this refactor (that hopefully works properly with this commit) I hope that it's now way clearer and easier to use and I can reuse it in other parts of the program without making errors. It's slightly less efficient because we're not reusing some of the results we reused before the refactor but that should be worth it.