Played around with performance analysis using os_signpost and added docs for some of the results. Minor refactoring.
The performance diff between the new advanced remapping engine and the old system is actually pretty minor when running an optimized build. The old one had 0.3 - 0.4% cpu usage when spamming a button mapped to "Back", and the new one has 0.4 - 0.6%.
If I understand correctly, profiler says that 50% of our usage is just inevitable as it's caused by system libraries doing stuff related to eventTaps, and we only have one eventTap for button input which we need so there's nothing we can do there I think.
Of the 50% we control, between 1/3 and 2/3 seems to be, because the way we retrieve active modifiers is pretty inefficient. (We call the function for getting active modifiers 9 times per button click and it's relatively slow)

So I think at best we could bring cpu usage when spamming a button down around 30% from 0.6% to 0.4%. So back to the old levels of performance. That's not really worth it I think. I might look into it more later.