We should almost definitely revert this change and come up with a better solution.
Made the helper not terminate itself after accessibility is enabled.

This is probably bad and leads to other issues.

I did this as a hotfix because displaying a notification in main app about which buttons are captured after accessibility has been enabled didn't work. It seems that launchd took too long to restart the helper after it killed itself and so it couldn't couldn't provide the main app with the set of captured button it requested. So in this change we have it not kill itself and just try to reconfigure itself from running in "ask for accessibility mode" to normal operation.