Added mechanism to heuristically figure out which devices activeModifiers' to feed to `reactToModifierChange()` when the change is triggered by pressing or releasing a keyboard modifier key. We're using the first device we find that has any pressed buttons. This heuristic should work really well in all realistic scenarios
Before we just called `reactToModifierChange()` once for each attached device, which caused issues in some edge cases, where both devices' modiferState stood in conflict when it comes to configuring the ModifiedDrag classes' state, which is global.

We initially planned to implement a cleaner solution for handling these conflicts, by having a separate ModifiedDrag state for each attached device.
However, this wouldn't have worked without some major refactoring, because we use an eventTap in the ModifiedDrag class to receive mouseDragged input. Unfortunately, eventTaps don't contain any infomation about which device caused an input, so we have no easy way of relating the MouseDraggedInput back to the device causing it to then obtain the device specific ModifiedDrag state. We'd have to build some hacks to obtain information about the device causing the input to make this work, and this would also be significantly slower than the current implementation.
So we decided to stick with a global ModifiedDrag state and use the heuristics explained above to still make it work as expected in most edge cases.