The main app is now checking whether the helper that it finds running is installed in a prefpane. If so, it removes that old helper from launchd and kills it. It also doesn't check the 'Enable Mac Mouse Fix' checkbox in this case.
With this change as well as the one from the last commit, updating from the old prefpane should be much less prone to issues.

I we can also handle App Translocation gracefully and maybe guide the user around the "App is from unidentified developer" messages better, we'd have a great installing experience I think.
Oh and we should also guide the user around the situations where the helper doesn't show up in the accessibility list, or when you have to uncheck and recheck it for it to work after an update (does this still happen under big sur?)