Tried to make capturing of keyboard shortcuts work, but it's wayyy harder than I anticipated...
I decided I want the popup button to be replaced with a text field when capturing shortcuts. That would be reminiscent of the System Preferences keyboard shortcut capturing UI.

But it's been a little hacky to get it to do anything. I just managed to make the cursor blink and I don't even know how.
We'd like the field to look like it's editable but we want to override the keydown behavour. Apparently this is not possible using an NSTextField without hacks so we'll probably have to switch to an NSTextView.