Tons of changes trying to get the keyboard capture fields to work. Their content can now be written to file and they work on a basic level, but they are extremely buggy and janky and also from a UX perspective the way they work now is not very good. Because when they hold a shortcut, they basically function as a shitty, weird popup button.
So in the future we're changing everything and having the shorcuts display in a popup button, and we'll only use the capture views for what they were named after - capturing input.

The horrible code I wrote for the current implementation doesn't work right anyways and this should make it a lot easier to make things work more reliably.