Input monitoring for keyCapturing is now using a cgevent tap, so that we can also capture system shortcuts like for spotlight. And prevent them from triggering while we are capturing.
We also delegated the input capturing to the helper app, because it already has accessiblity privileges that allow it to do that.

I'm not sure if the main app also has those privileges if the helper has them. It would make sense to me, since it's in the same bundle, but I didn't test that because I thought probably not.

If it does turn out that we don't need the helper for this we should refactor this, but for now it seems fine.