Added new mechanisms for detecting if an old (or another) version of the helper app is registered with launchd and then cleans that old stuff.
I hope this will prevent some issues that people had when upgrading from the old prefpane version to the new app  version.

From the few reports I've gotten, it seemed like there was some sort of conflict with the old and the new version both being registered with launchd. I couldn't reproduce the issues, but I think this has a pretty good chance of fixing that.

If we then also implement automatic untranslocation of the app, I hope that will make upgrading very solid.

For reports, see:
- The end of this gh issue thread: https://github.com/noah-nuebling/mac-mouse-fix/issues/54#issuecomment-811482406
- The mail by Jeff starting with "I attached the system log."