Lots of refactoring. Might have broken stuff.
Implemented dynamic checking for pre-release or regular update feed based on settings.

Implemented a system to check whether a launch of an application is the first launch of that version or the first launch of the app in general. Tested that a little and it seems to work.

Also trying to test if the app has been launched by Sparkle Updater. Not sure if that works.

If the app has not been launched by Sparkle Updater and it's the first launch of that version, then we override the checkForPreReleases setting.