Fixed issue where simulated click and drag wouldn't work.
There are issues with fast user switching. See https://github.com/noah-nuebling/mac-mouse-fix/issues/93.

To fix these issues we changed all event taps to use the kCGSessionEventTap tapLocation.

But that introduces tons of new errors, because when we post events at the kCGSessionEventTap tapLocation (which we always do when we post events) they'll get re-picked up by the event taps at that location. This is really bad and could lead to tons of issues. It's a wonder that it works as decently as it does from my quick tests but I'm sure it introduces lots more subversive issues.

One of these issues is the one we fixed in this commit.