Added toast notifications saying "Mac Mouse Fix will stay enabled when you restart your Mac" after enabling Mac Mouse Fix but disabled them again.
Added a tooltop saying "Mac Mouse Fix will stay enabled when you restart your Mac" though

It worked fine for the most part but the problem is that we want to display this when the Accessibility Overlay disappears. After the overlay disappears, we want to display both the "will stay enabled" notifications AND the "capture" notifications.

Displaying both of this doesn't work properly.

Currently, when a notification is attached while an old one is still displaying, the old one is dismissed. This is what we want in most cases, but here it poses a problem.

Ideas for solution
- Implement "onClose" block for notifications, and use such a block to display the next notification → This seems relatively simple and flexible.
- Create a queue of notifications, where the next one appears after the previous one disappears, and then be able to choose whether we want to display a notification after the previous one is dismissed or immediately
- Display all the info in one big notification.
     - We'd have to change how our CaptureNotifications utility class works. Such that capture notifications aren't displayed directly by it, but instead it returns a string which we can append stuff to. → Shouldn't be too hard
- Use normal modal alerts instead of the toast notifications, (not sure if that would make things easier to implement)
- Use a popover for the 'will stay enabled' tip
- Display the second alert after a hardcoded delay