Manually changed objectVersion in the pbxproj file from 54 to 48.
This will hopefully make the project compatible with older versions of Xcode until version 8.0.

Davina told me in an email that she couldn't open the 2.0.0 source code from Xcode 11.3.1, because the project was set to only be compatible with newer versions of Xcode.

However the source code from the master branch (1.0.0) still worked for her for some reason.

Normally this error can be fixed by changing the project format in the Project inspector, but in my case it didn't. The Project format was set to "Xcode 8.0", but it didn't work on Xcode 11.

According to this StackOverflow thread (https://stackoverflow.com/questions/49622341/failed-to-load-project-incompatible-project-version-pop-up-appears-when-i-op)
changing the Project Format in Xcode works the same as changing the
`compatibilityVersion` and `objectVersion` values in the pbxproj file. But when I did, only the `compatibilityVersion` changed. So in this commit I manually changed the `objectVersion` to the value it should be for a project format of "Xcode 8.0" (according to the SO thread mentioned above).

I hope this doesn't break anything.
Maybe there was a good reason that the objectVersion wasn't changed when setting the Project Format from within Xcode?

Edit: I just found a possible reason for the objectVersion being set higher than it should be given the Project Format / compatibilityVersion:

All of the objects in my xib files have "File Inspector > Interface Builder Document > Document Editing > Opens in" set to "Latest Xcode (12.0)"
I tried to set this to 11.0, but it's reset to 12.0 whenever I open the file again unfortunately.

So I'll just let Xcode do what it does and let it set the objectVersion to 54, making it only compatible with Xcode 12 and later. (In the next commit)