Changed Project Format to 9.0 and then back to 8.0. This made Xcode reset the objectVersion to 54 making the project only compatible with Xcode 12 and higher.
See previous commit for context.

Edit: I just tested the project from master. When I set the Project Format to 9.3 and then back to 8.0 there, the objectFormat will be reset to 48 (?!).
The xib files also have "Opens in" set to version 12.0 on master but it still works.

So I have no idea why it doesn't set it to 48 on the feature-remap branch, but I assume it has some reason, and it will probably set it to 54 automatically again at some point anyways, so I won't touch it.