Added Carthage derived data folder to gitignore, and cleaned up project structure a bit.
My goal was to integrate ReactiveObjc using Carthage, but that doesn't seem to work.
The problem seems to be that it tries to build all build schemes for all platforms including watchOS and so on. Despite the fact that I specify the platform to be macOS (I used this command specifically `carthage update --use-xcframeworks --platform macos`)

Since I deleted the watchOS and tvOS sdks from Xcode, these build schemes will fail, leading all of `carthage update` to fail. I can't find a way for it to only build the macOS version. I'll look into incooperation CocoaPods now instead.