Swift implementation of BezierCurve works now.
We already had an ObjC implementation for an animation curve, that is a cubic BezierCurve (cubic <=> 4 control Points <=> degree 3) with the first and last control points fixed to (0,0) and (1,1) (Find it in AnimationCurve.m). We copied that from some Apple source code. I think WebKit, but I can't find it anymore.

The new Swift implementation features these improvements over the Apple implementation:
- It supports an arbitrary number of control points. Not just 4.
- The position of these control points is arbitrary, as well. The start and endpoints arent fixed.

This should make it a little easier to deal with for creating acceleration curves, which is why we wrote this.

However it was probably not worth the hassle, as it would've been sufficient for our needs and much easier would we have just written a wrapper around the ObjC class which would allow us to easily scale and translate the curve along the x and y axis → That would make it possible to set the start and end control points to arbitrary locations using the ObjcC implementation under the hood.
Having more that 4 control points is probably unnecessary to begin with. All the online editors have cubic beziers with first and last points (0,0) and (1,1). And path tools in graphic design programs usually just stick together lots of cubic bezier curves.
Also my Swift implementation is slower than the AnimationCurve.m implementation, because they somehow solve the Bezier Curve and turn it into a polynomial while I use De Casteljau's algorithm. But it should still be way fast.

So yeah that was a waste of time. But at least it was kinda fun!