Turned DerivedProperties into a class instead of an extension.
That way it might work with ObjC but it won't work with Value types anymore.

Also this makes it possible to declare derived properties inline and not in the init function, making them usable for static properties. (Maybe that was possible before, and I didn't get how - will check now)