Tested how the MX Master sends events.
It never sends events that aren't perpendicular to an axis. Event when you scroll on the horizontal and vertical mouse wheels at the same time.

So our idea, where we determinte the final scrolling direction (with user settings and modifiers applied to device input on the horizontal or vertical axis) all in one centralized place right when the events come in, and then do all of our animation stuff only in one direction, and then apply the direction again, when sending the events, should work just fine.
So all the nitty gritty code only has to deal with one direction.