Upgraded pixelated animator to use biased subpixelator so it works the same with negative value ranges which it's now used with for the momentum scroll code in GestureScrollSimulator.m
Also I noticed that the momentum scrolling now sometimes doesn't work at all in Safari. I don't think that happened before we started using the pixelated animator. Other then that the pixelated animator doesn't seem to help with anything at least not that I've noticed (there are lots of little quirks with how different programs interact with the scrolling so it might very well be that there's some other place where it does help) But in any case this is a good opportunity to debug the pixelated animator further.