Compared events which the pixelatedAnimator sends through the GestureScrollSimulator to the events sent by the trackpad. In hopes of fixing issue where single scroll tick won't scroll at all in Safari.
I tried to send a kIOHIDEventPhaseMayBegin phase right before the kIOHIDEventPhaseBegan phase event, because the trackpad will do that sometimes, but it didn't help.

Also tried to use a rounding subpixelator instead of a biased one for the line deltas, because the first line delta is sometimes 0 on the trackpad, but that didn't help, either.

I can't see any other differences between the events we're sending and the trackpad events, at least in the way we're driving our  [postGestureScrollEventWithGestureVector:] function. So maybe that function is the culprit, or the issue is in Safari and it's just not as noticable using a trackpad, because it's almost impossible to scroll a single pixel on the trackpad.

Edit: After doing some more testing with Safari, I'm fairly certain the same issue occurs when using the trackpad - it's just not as noticable.