Fixed issue where scroll direction was inverted for single ticks.
The issue was caused by Newtons method in Bezier.swift finding a diffent t outside of the range [0,1] for which x(t) equals the x we're searching for.

I fixed it by modifying Newtons method in Bezier.swift to always keep t inside the [0,1] interval. Not sure if this has bad sideeffects. It doesn't seem to.