We were thinking about improving the modifier system to allow for combining modifiers, as well as for modifying scrolling to produce zoom events, horizontal scrolls, etc.
In the process of this I took a closer look at ButtonTriggerHandler and ButtonLandscapeAssessor.

While doing that I fjound that ButtonLandscapeAssessor should still work, if we swap out the method by which we apply modifier-based overrides to the remap dict (which is necessary to get combinations of modifiers working.)

I also documented and cleaned up the ButtonLandscapeAssessor class and others a bit.

I assume that the ButtonTriggerHandler is the only other part of the system that uses the overriding mechanism that we would change, and my feeling is that it would still work as well. I feel like I engineered everything in anticipation of this.

Next thing to think about is: how do we compartmentalize (and possibly cache) the new modfier overriding of the remap dict? We need to use it for applying modifiers to
- buttons
- drag
- scroll

And when we recalculate the override everytime we use it, that might be pretty slow.

But then again, don't optimize prematurely, but make sure you have a logical model for how this shared resource is accessed that makes some sense.