Further cleanup of ButtonLandscapeAssessor, ButtonTriggerHandler and ModifierManager.
Have a pretty good idea of how to go forward now.

I will create a separate class for the + (MFEffectiveRemapsMethod)effectiveRemapsMethod_Override function and I'll rename it. (this function combines different remappings for different modificationPreconditions together based on which modifiers are pressed. ATM it's pretty primitive, but we'll extend it later.) This function is used by ModifierManager to know when to activate ModifiedDrags, and by ButtonTriggerHandler to figure out what actions to fire when, and it will be used by our rewritten version of ScrollModifiers to implement ModifiedScroll.

After that we will  rewrite ScrollModifiers to do new stuff. Specifically we want a getModifications() function which returns a struct with 2 fields "inputModification" which can contain fastScroll, preciseScroll, or none, and another field "effectModification" which can contain magnification, horizontal, rotate, none, and maybe others in the future.

It will fill this struct based on the overriden remaps dict it gets from the class which we based around the effectiveRemapsMethod_Override function.

I don't think we need the event sending and keyboardModifier monitoring capabilities in this class anymore. That's primarily because we don't need to generate events with start and end phases based on when the keyboard modifiers are pressed anymore, because we can use the animationPhases of our new scrolling System to drive the touchEventPhases.

Then we need to define some new structure for the remap dict to define the different scroll modifications, and set it up in such a way that the modifications of the different slots (effect and input modification) won't override each other when two different modification dicts are overriden into each other. (That's what effectiveRemapsMethod_Override does)
Or maybe we can custom-taylor our future overriding algo to make sure values in the 2 slots can coexist.

After this we should be able to get ModifiedScroll preconditioned on holding mouse buttons working!

Then, once this is imlemented, we can extend the effectiveRemapsMethod_Override() function to allow for modifier-key-combinations. I thougth this through and it will be relatively involved on the logic side. If it also turns out to be slow (probably not) we can look into caching the results between different calls or other optimizations.