Lots of changes. Primarily centered around new `puppetCursor` mechanism in modifiedDrag.
During a twoFingerSwipeModifiedDrag (gestureScroll), we wan the point that the scrolls are dispatched to to stay the same while the user moves the mouse (you don't accidentally move the dispatch point outside of the window that you're scrolling on and it stops scrolling)
But you also don't want the cursor to freeze while using twoFingerSwipeModifiedDrag, because that forces the user to eventually lift off the mouse off the table and reposition, because the cursor will drift down if you're scrolling down, left if you're scrolling left, etc.

This is a real problem because I couldn't find a trivial way to decouple the scroll dispatch point and the cursor position.

But now we've done it by drawing a fake image of the cursor on a transparent overlay and hiding the real cursor which just serves as the stable dispatch point for scroll events, and a bunch of other hacks!
And it's all so seemless and awesome that the user doesn't even notice this whole trickery!

The only thing I don't love about the twoFingerSwipeModifiedDrag is that the initial momentumScroll speeds are too low for slow speeds and too fast for fast speeds.

But other than that it's greattt!