Built and refined all the "Click and Scroll" features. Also built the UI for them.
Even built an unplanned one "Desktop & Launchpad" which I really like!
I didn't plan to build it because I thought "Using click and click and hold for Desktop and Launchpad respectively is already an option and probably much better and easier to use than click and scroll" but it just feels so good! And also it feels similar to Click and Scroll to zoom in & out which makes it really easy to learn and even more intuitive when you have both set up.