Fixed some more concurrency issues.
The twoFingerModifiedDrag deactivation function now waits for momentumScroll to start before warping the pointer back, which fixes a bunch of race conditions. This was really hard and messy to implement, but it works!

The pointer is still a little laggy during twoFingerModifiedDrag, and sometimes pointer movement is swallowed during it which sucks, but I don't know if I can make it any better.