Cleaned up inputIsCursorMovement flag and other stuff.
Back in the day we experimented with stopping cursor movement by 'seizing' the HID device, which prevents the OS from reading any input from the mouse.

This was super buggy and unstable though. We definitely won't use it. We've found better ways of freezing the pointer.

There is still a lot of code in Device.m that was for this purpose, and is now unused. We should remove it too at some point.