Fixed bug where Helper would crash when using Swift Scroll after 360* Scroll. Still a lot of bugs in their interaction.
I'm thinking about implementing a mechanism in GestureScrollSimulator which detects when a new driver tries to drive it and then shuts down the old one ASAP before handling the requests of the new driver. This should fix a lot of issues with how 360* Scroll interacts with Scroll Wheel scrolling.

(The two drivers for GestureScrollSimulator are Scroll.m and ModifiedDrag.m)