Added a class GlobalEventTapThread which provides an extra thread to run eventTaps on besides the main thread.
We're using this for the mouseMoved/Dragged callbacks which are used for twoFingerModifiedDrag. It does slightly improve performance.

But twoFingerModifiedDrag is still more than 2x slower compared to threeFingerModifiedDrag. Don't know what's going on. Instruments says that 'Library functions' take up like 70% of the CPU time. It's weird.

This commit also includes clean up and other improvements.

Also we've found that just having an eventTap that does nothing except dispatch a block to another queue is very very slow. When we do fast wiggling of the mouse with twoFingerModifiedDrag active, that takes around 6% CPU in Activity monitor on my M1 mac. One event tap doing nothing takes 1%. Since we use two separate event taps for PointerFreeze and ModifiedDrag, removing one of those eventTaps is a pretty good optimization opportunity. But we should do that once we create our own abstraction around eventTaps.