Experiments with open source IOKit headers and private functions.
Managed to get an "HIDEvent" from CGEventRef. "HIDEvent" is not opaque and much easier to work with! Also, "HIDEvent" has a "senderID" property which is the IORegistryEntry of the sending device!!! This means we don't have to use Use IOHIDManager and then setup callbacks for matched IOHIDDevice and do this weird communication between the IOHID callbacks and the CGEventTaps in order to know which device sent an event in the eventTap!!!! This is huge. Will greatly simplify and device specific settings or support for device specific features.