Some cleanup and docs.
Notes:

We built a new analysis project for HIDEvent. We can't recreate Force touch unfortunately. It doesn't even work in our event replay. Everything else works in our event replay. So it seems like there is some black magic involved with creating working force touch events. Or I'm overlooking something.

I also decided not to recreate the gesture scrolls as HIDEvents (Still a little on the fence actually). It would be too much work and the current implementation is almost perfect and I think it's shippable - so we shouldn't priioritize rebuilding it for now.

Was interesting to look at how HIDEvent works and all the HIDEventFields though. This would've made reverse engineering the touch stuff so much easier! but we did it anyway eyy.