Fixed duration-based init for DragCurve. Found great 'inertial' setting for scroll algorithm. Added file for notes on setttings I found.
The inertial setting actually uses the same Drag parameters we use to to approximate the Apple Trackpad scrolling algorithm. This means we can actually send these events as momentum scrolls and get better behaviour for scroll bouncing and other things!!

Will build mechanism that can send the Drag part of the BezierHybridCurve as momentum scrolls.