Todo to finish up new scrolling system:
- Polish the dynamic scroll config. Like adjusting the animation duration when using scroll to zoom
- Implement formular that calculates pxPerTickEnd based on semantic scrollAcceleration value that the user can choose as well as screen real-estate and inertia level (See ScrollConfigTesting for this formula)
- Think about also changing pxPerTickBase dynamically based on screen canvas size and inertia level as well
- Fix some bugs
	- Bug where sometimes cursor disappears after using Drag Scroll when the computer is slow → Make sure the cursor stuff is EXTREMELY robust. Just crash the app if anything only comes close to going wrong.
- Final polish for interaction between different modificaitons (e.g. Starting to drag while you're clicking-and-scrolling)
- Build the scroll settings UI and connect it up with the backend

In this commit:

Refactored ScrollConfig such that it is copyable and has a singleton master instance that only changes when the config file changes. This allows us to cache most values in the singleton master instance even when we dynamically override some config values depending on the currently active modifiers in Scroll.m.

In practise, this is actually slightly slower than the previous implementation. I think this is because the copying forces all the lazy computed variables to be computed or something? But that doesn't make sense they should only be computed once in the master instance and then copied over? I don't understand this.

Edit: Tried to analyze what's going on using Instruments Time Profiler and it didn't help me understand things better. The new scrolling code is soo much slower than the old MMF2 implementation, but in the profiler none of my functions shows up with more than like 15%. And the bulk is the event sending functions which I can't really optimize I think?? I suspect that maybe some stuff with the mutex locks waiting on each other makes things slower than they need to be. Or maybe there is something going wrong with the lazyness and the copying. I think I had an idea for this but I forgot.