Found bug where stopping momentumScroll doesn't work anymore from Scroll.m.
The 2 concurrency changes we made recently in DisplayLink and VectorAnimator, both independently caused this issue.
The 2 changes were:
1. Putting VectorAnimator on the same concurrency queue as DisplayLink.
	→ We were async dispatching to the displayLinkQueue again when calling displayLink.start() and displayLink.stop() from VectorAnimator and already running on the displayLinkQueue. This caused the order of operations to be messed up. We just forgot to change this when we combined the 2 dispatch queues. Now it works fine!
2. Async dispatching to mainQueue to call CVDisplayLinkStart() / CVDisplayLinkStop() from DisplayLink.startI() and DisplayLink.stop().
	→ This directly breaks the isRunning() function. We reverted to sync dispatching for now in hopes that other changes we made fixed the deadlocks, but if not we also have another idead for a solution that we documented in code.

God this multithreaded code is so dangerous, I feel like it's so close to working perfectly but still so far. It's like 99.9% percent of the time it works perfectly but in those 0.01 percent it breaks it breaks in super confusing and always different ways, and when I fix an issue it introduces 3 new ones that I don't notice until later.

This is the first time where I feel like automated tests are beginning to make sense.

Idk why I put so much more effort and detail into this commit message than usually