Finally have the Scrolling screen working.
Took forever because I wanted to make the ConfigValue class more generic so it works with NSEvent.ModifierFlags as well. But swift is tooo stupid.
So I gave up on that which is good because I had spent over an hour on that and it's not worth it just for "pretty code"