Not sure if this commit makes sense.
Refactored Objects.m because it crashed when running the helper standalone (outside the mainApp bundle).

It didn't use to crash on previous macOS versions I'm pretty sure. But maybe we should crash, because we're normally never running the helper outside the mainApp bundle.

Also we might have broke some things in the refactor.

So not sure if this commit makes sense.