Deleted Accomplice and cleaned up HelperServices in the same commit.
This is bad because I'm not sure I really want to remove Accomplice. The removal might break stuff when relocating MMF while it's running. If it does, add Accomplice back in.

Edit: Doing some superficial testing on Ventura, it didn't break things to remove Accomplice. But it also doesn't make the app any smaller or compile times any shorter, so questionable if worth it.

TODO: Test if relocating is broken now on macOS 12 (if no, do thorough testing on macOS 13)