Major refactor of the licensing code
Most important changes are that

- LicenseSheet doesn't use the Gumroad class directly anymore, but instead uses License class.
- Gumroad class is fileprivate now and can only be interacted with through License class.
- Also the nOfActivations validation code has been moved from Gumroad class to License class.
- Lots more small cleanup and renaming, resturucturing, documentation added.

This shouldn't change anything functionally, but should make future changes and enhancements easier due to much clearer structure and documentation.

Well when activating a license on the licenseSheet there is a bunch of extra stuff being done. Maybe there are some edge cases which behave differently now but I couldn't think of anything specific.

As far as bugs I might have introduced there is a lot of potential for that, but in my basic tests everything still seems to be functioning great. I tested trying to activate without internet connection and trying to activate a license with too many activations. Both behave as expected.