License.swift now supports free countries!
I'll just have to give proper UI feedback about this on the about tab and when trying to activate a license, then I should be done with this feature.

Also had to do some refactoring and restructuring to make this work. Notably I abstracted away caching for no internet in License.swift.
And I restructured the logic for overriding the core code's decision about whether a license is valid or not and then passing up the reason for the override and caching the reason for when there is no internet.

Pretty big potential of mistakes here.